prevent admin from being removed

+++ craftboard/account-delete.php
@@ -2,6 +2,9 @@
     require 'account-common.php';
     require 'database.php';
     require 'adminrestrict.php';
+    if ($_POST['username'] == "admin") {
+        exit();
+    }
     $query = $database->exec('DELETE FROM users WHERE username="'.$_POST['username'].'"');
     header('Location: ' . $_SERVER['HTTP_REFERER']);
 ?>
