+++ craftboard/settings.php
@@ -85,7 +85,12 @@
                             $sql = $database->prepare('SELECT username FROM users');
                             $result = $sql->execute();
                             while ($data = $result->fetchArray()) {
-                                    echo '<tr><td>'.$data['username'].'<a onclick="return confirm(\''.$lang['areyousure'].'\')" href="account-delete.php?username=' . $data['username'] . '"><button type="button" class="btn btn-sm btn-danger m-2"><i class="fa-solid fa-trash"></i></button></a></td></tr>';
+                                    if ($data['username' == "admin"]) {
+                                        echo '<tr><td>'.$data['username'].'</td></tr>';
+                                    }
+                                    else {
+                                        echo '<tr><td>'.$data['username'].'<a onclick="return confirm(\''.$lang['areyousure'].'\')" href="account-delete.php?username=' . $data['username'] . '"><button type="button" class="btn btn-sm btn-danger m-2"><i class="fa-solid fa-trash"></i></button></a></td></tr>';
+                                    }
                             }
                            ?>
                            </tbody>
